+++ src/pages/home/Posts.tsx
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+import PostCard from "../../components/PostCard/PostCard";
+import { useAppSelector } from "../../hooks";
+
+export default function Posts(): JSX.Element {
+  const { postsOfFollowingUsers } = useAppSelector((s) => s.postsReducer);
+
+  return (
+    <Container>
+      {postsOfFollowingUsers?.map((post) => (
+        <PostCard key={post?.pid} post={post} />
+      ))}
+    </Container>
+  );
+}
+
+const Container = styled.section`
+  display: flex;
+  flex-direction: column;
+`;
